Reorganized directory
- Reorganized directory so that the coding exercise files are not at the top level
- Will want to maintain their function in the web app but also now add robust features for the fantasy shop
- Ideally there will be a completely different looking website for the two
