+++ package.json
@@ -3,11 +3,11 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "lint": "ng lint"
+    "start": "nx serve",
+    "build": "nx build",
+    "watch": "nx build --watch --configuration development",
+    "test": "nx test",
+    "lint": "nx lint"
   },
   "private": true,
   "dependencies": {
@@ -30,6 +30,8 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^16.1.7",
+    "@angular-devkit/core": "~16.1.7",
+    "@angular-devkit/schematics": "~16.1.7",
     "@angular-eslint/builder": "16.1.0",
     "@angular-eslint/eslint-plugin": "16.1.0",
     "@angular-eslint/eslint-plugin-template": "16.1.0",
@@ -37,6 +39,8 @@
     "@angular-eslint/template-parser": "16.1.0",
     "@angular/cli": "~16.1.7",
     "@angular/compiler-cli": "^16.1.0",
+    "@nx/angular": "16.5.1",
+    "@nx/workspace": "16.5.1",
     "@schematics/angular": "^16.2.0",
     "@types/jasmine": "~4.3.0",
     "@typescript-eslint/eslint-plugin": "5.62.0",
@@ -48,6 +52,8 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
+    "nx": "16.5.1",
+    "nx-cloud": "latest",
     "typescript": "~5.1.3"
   }
 }
