+++ src/app/coding-exercises/fibonacci/fibonacci-number-generator/fibonacci-number-generator.component.scss
