+++ src/app/coding-exercises/fibonacci/fibonacci-sequence-generator/fibonacci-sequence-generator.component.spec.ts
