+++ src/app/coding-exercises/fibonacci/fibonacci.component.scss
@@ -1,4 +1,4 @@
-@import '../../styles.scss';
+@import '../../../styles.scss';
 
 
 .wrapper {
