+++ src/app/coding-exercises/reverse-int/reverse-int.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { ReverseIntService } from '../reverse-int.service';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { ReverseIntService } from '../fibonacci/services/reverse-int.service';
 
 @Component({
   selector: 'app-reverse-int',
