+++ src/app/coding-exercises/reverse-int/reverse-int.module.ts
